requirements.txt
# UVPFL - User Profile-Based Viewport Prediction Using Federated Learning
# Requirements for the implementation

# Deep Learning (GPU)
tensorflow[and-cuda]>=2.15.0

# Data Processing
numpy>=1.23.0
pandas>=1.5.0
scipy>=1.9.0

# Computer Vision
opencv-python>=4.6.0
Pillow>=9.0.0

# Machine Learning
scikit-learn>=1.1.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.64.0

# Jupyter support (optional)
jupyter>=1.0.0
ipykernel>=6.0.0
